refactor: decompose torrent and dht modules to comply with 400-line maintainability limit (Issue #96)
