+++ aura-docs/project/TASKS.md
@@ -7,7 +7,7 @@ This document tracks the technical debt and missing features identified during t
     - Implement `fallocate` for actual block allocation (Linux).
     - Implement COW-awareness (`chattr +C`) for Btrfs/ZFS.
     - Implement Windows long path prefixing (`\\?\`).
-- [ ] **Modular Architecture Refactor** (Issue #96)
+- [x] **Modular Architecture Refactor** (Issue #96)
     - Decompose `aura-core/src/torrent.rs` (461 lines) into `torrent/`.
     - Decompose `aura-core/src/dht/mod.rs` (445 lines).
